Start porting box_easy test
This doesn't have the same output as you'd find at:

https://github.com/jedisct1/libsodium/blob/master/test/default/box_easy.exp

I don't know why, but I'm confident that I'm doing something wrong. I'm
very concerned about why my test has `c` with a length of 146 (which is
the length of the message plus 16 bytes for the HMAC) but the original
test has an *extra* 16 bytes... for another HMAC? I don't know.