Fix backward crypto_secretbox_detached bug
Problem: Both `crypto_secretbox_detached` and the corresponding
`crypto_secretbox_open_detached` have a bug where the MAC is added as a
suffix rather than the prefix. This creates a problem where the methods
are compatible with each other but incompatible with other libsodium
implementations.

Solution: Reverse the backward implementation and ensure that the MAC is
added to the output as a prefix rather than as a suffix.