Remove unused argument from crypto_kx
Problem: randombytes_buf uses the size of the buffer as the number of
bytes to output, so we don't need to add an argument about the number of
bytes we want.

Solution: Remove unused argument and use buffer size assertion to be
sure that we're producing the correct number of random bytes.