Remove unused argument from crypto_kdf
Problem: We're passing an extra argument, which looks like it's using
the method signature for `TypedArray.prototype.subarray()`, which gives
you the option of setting the end of the array. Since this method
doesn't give us an optional third parameter the argument is being
ignored.

Solution: Remove the unused argument.