BREAKING: Rename '_ready' suffix to '_wasm_ready'
When looking through each of the extra methods exposed by this module I
noticed lots of 'wasm' mentions, but since this method doesn't mention
wasm it wasn't clear to me why this extra method was being exposed.

This commit is a breaking change that renames crypto_generichash_ready
to crypto_generichash_wasm_ready, which feels more clear to me.